shifts more common functionality to base forecaster
